requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pymongo==4.15.2
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
bcrypt==4.3.0
PyJWT==2.10.1
python-multipart==0.0.12
email-validator==2.3.0
tencentcloud-sdk-python==3.0.1481
cos_python_sdk_v5==1.9.38
motor==3.7.1
passlib==1.7.4
python-jose==3.5.0
cryptography==44.0.0
starlette==0.48.0
websockets==13.1
requests==2.32.5
pyasn1==0.6.1
pyasn1-modules==0.4.2

